refactor: move resource rate logic to selector
