+++ src/components/useResourceSections.js
@@ -1,117 +1,24 @@
 import { useMemo } from 'react';
-import { computeRoleBonuses } from '../engine/settlers.js';
-import { RESOURCE_LIST } from '../data/resources.js';
 import { RADIO_BASE_SECONDS } from '../data/settlement.js';
-import { formatRate } from '../utils/format.js';
 import {
-  getCapacity,
-  getResourceRates,
+  getResourceSections as getResourceSectionsSelector,
   getSettlerCapacity,
 } from '../state/selectors.js';
-
-const CATEGORY_LABELS = {
-  FOOD: 'Food',
-  RAW: 'Raw Materials',
-  SOCIETY: 'Science',
-  CONSTRUCTION_MATERIALS: 'Construction Materials',
-  ENERGY: 'Energy',
-};
 
 export function useResourceSections(state) {
   const settlers = state.population?.settlers?.filter((s) => !s.isDead) || [];
 
-  const roleBonuses = useMemo(() => computeRoleBonuses(settlers), [settlers]);
-
-  const netRates = useMemo(
-    () => getResourceRates(state, true, roleBonuses),
+  const resourceSections = useMemo(
+    () => getResourceSectionsSelector(state),
     [
       state.resources,
       state.buildings,
       state.research?.completed,
       state.population?.settlers,
       state.gameTime,
-      roleBonuses,
+      state.foodPool,
     ],
   );
-
-  const prodRates = useMemo(
-    () => getResourceRates(state, false, roleBonuses),
-    [
-      state.resources,
-      state.buildings,
-      state.research?.completed,
-      state.gameTime,
-      roleBonuses,
-    ],
-  );
-
-  const entries = useMemo(() => {
-    const groups = {};
-    const foodIds = [];
-    RESOURCE_LIST.forEach((r) => {
-      if (
-        r.id === 'power' &&
-        !(state.research.completed || []).includes('basicEnergy')
-      )
-        return;
-      const amount = state.resources[r.id]?.amount || 0;
-      const isFood = r.category === 'FOOD';
-      const capacity = isFood ? null : getCapacity(state, r.id);
-      if (isFood) foodIds.push(r.id);
-      const rateInfo = (isFood ? prodRates : netRates)[r.id];
-      const discovered = state.resources[r.id]?.discovered;
-      if (rateInfo.perSec !== 0 || amount > 0 || discovered) {
-        if (!groups[r.category]) groups[r.category] = [];
-        groups[r.category].push({
-          id: r.id,
-          name: r.name,
-          icon: r.icon,
-          amount,
-          capacity,
-          rate: rateInfo.label,
-          tooltip:
-            r.id === 'power'
-              ? 'Power is generated by Energy buildings and stored in Batteries. Excess above capacity is lost.'
-              : undefined,
-        });
-      }
-    });
-
-    if (groups.FOOD) {
-      const totalAmount = foodIds.reduce(
-        (sum, id) => sum + (state.resources[id]?.amount || 0),
-        0,
-      );
-      const totalCapacity =
-        state.foodPool?.capacity ??
-        foodIds.reduce((sum, id) => sum + getCapacity(state, id), 0);
-      const totalNetRate = foodIds.reduce(
-        (sum, id) => sum + (netRates[id]?.perSec || 0),
-        0,
-      );
-      const totalRow = {
-        id: 'food-total',
-        name: 'Total',
-        amount: state.foodPool?.amount ?? totalAmount,
-        capacity: totalCapacity,
-        rate: formatRate(totalNetRate),
-      };
-      groups.FOOD = [totalRow, ...groups.FOOD];
-    }
-
-    return Object.entries(groups).map(([cat, items]) => ({
-      title: CATEGORY_LABELS[cat] || cat,
-      items,
-      defaultOpen: true,
-    }));
-  }, [
-    state.resources,
-    state.buildings,
-    state.research?.completed,
-    state.foodPool,
-    netRates,
-    prodRates,
-  ]);
 
   const totalSettlers = settlers.length;
   const capacity = getSettlerCapacity(state);
@@ -146,7 +53,7 @@ export function useResourceSections(state) {
 
   const sections = useMemo(() => {
     const rendered = [];
-    entries.forEach((g) => {
+    resourceSections.forEach((g) => {
       rendered.push(g);
       if (hasRadioResearch && g.title === 'Science')
         rendered.push({ title: 'Settlers', settlers: true });
@@ -155,7 +62,7 @@ export function useResourceSections(state) {
       rendered.push({ title: 'Settlers', settlers: true });
     rendered.push({ title: 'Happiness', happiness: true });
     return rendered;
-  }, [entries, hasRadioResearch]);
+  }, [resourceSections, hasRadioResearch]);
 
   const settlersInfo = {
     total: totalSettlers,
