+++ src/state/selectors.js
@@ -2,6 +2,7 @@ import { BUILDINGS, PRODUCTION_BUILDINGS, getBuildingCost } from '../data/buildi
 import { RESOURCES } from '../data/resources.js';
 import { RESEARCH_MAP } from '../data/research.js';
 import { getSeason, getSeasonMultiplier } from '../engine/time.js';
+import { computeRoleBonuses } from '../engine/settlers.js';
 import { formatRate } from '../utils/format.js';
 import { BALANCE } from '../data/balance.js';
 import { ROLE_BY_RESOURCE } from '../data/roles.js';
@@ -175,4 +176,78 @@ export function getResearchStorageBonus(state, resId) {
     if (effectApplies(e, resId)) bonus += e.percent || 0;
   });
   return bonus;
+}
+
+const CATEGORY_LABELS = {
+  FOOD: 'Food',
+  RAW: 'Raw Materials',
+  SOCIETY: 'Science',
+  CONSTRUCTION_MATERIALS: 'Construction Materials',
+  ENERGY: 'Energy',
+};
+
+export function getResourceSections(state) {
+  const settlers = state.population?.settlers?.filter((s) => !s.isDead) || [];
+  const roleBonuses = computeRoleBonuses(settlers);
+  const netRates = getResourceRates(state, true, roleBonuses);
+  const prodRates = getResourceRates(state, false, roleBonuses);
+
+  const groups = {};
+  const foodIds = [];
+  Object.values(RESOURCES).forEach((r) => {
+    if (
+      r.id === 'power' &&
+      !(state.research?.completed || []).includes('basicEnergy')
+    )
+      return;
+    const amount = state.resources[r.id]?.amount || 0;
+    const isFood = r.category === 'FOOD';
+    const capacity = isFood ? null : getCapacity(state, r.id);
+    if (isFood) foodIds.push(r.id);
+    const rateInfo = (isFood ? prodRates : netRates)[r.id];
+    const discovered = state.resources[r.id]?.discovered;
+    if (rateInfo.perSec !== 0 || amount > 0 || discovered) {
+      if (!groups[r.category]) groups[r.category] = [];
+      groups[r.category].push({
+        id: r.id,
+        name: r.name,
+        icon: r.icon,
+        amount,
+        capacity,
+        rate: rateInfo.label,
+        tooltip:
+          r.id === 'power'
+            ? 'Power is generated by Energy buildings and stored in Batteries. Excess above capacity is lost.'
+            : undefined,
+      });
+    }
+  });
+
+  if (groups.FOOD) {
+    const totalAmount = foodIds.reduce(
+      (sum, id) => sum + (state.resources[id]?.amount || 0),
+      0,
+    );
+    const totalCapacity =
+      state.foodPool?.capacity ??
+      foodIds.reduce((sum, id) => sum + getCapacity(state, id), 0);
+    const totalNetRate = foodIds.reduce(
+      (sum, id) => sum + (netRates[id]?.perSec || 0),
+      0,
+    );
+    const totalRow = {
+      id: 'food-total',
+      name: 'Total',
+      amount: state.foodPool?.amount ?? totalAmount,
+      capacity: totalCapacity,
+      rate: formatRate(totalNetRate),
+    };
+    groups.FOOD = [totalRow, ...groups.FOOD];
+  }
+
+  return Object.entries(groups).map(([cat, items]) => ({
+    title: CATEGORY_LABELS[cat] || cat,
+    items,
+    defaultOpen: true,
+  }));
 }
